requirements_mr_gates_new.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0   
anyio==4.12.0
attmap==0.13.2
attrs==25.4.0
babel==2.17.0
blobfile==3.1.0
certifi==2025.11.12      
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1      
comtypes==1.4.13
csvw==3.7.0
dataclasses-json==0.6.7
distro==1.9.0
dlinfo==2.0.0
eido==0.2.4
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.0
langchain-text-splitters==1.0.0
langsmith==0.4.59
language-tags==1.2.0
logmuse==0.2.8
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.3.5
ollama==0.6.1
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
openai==2.11.0
orjson==3.11.5
oyaml==1.0
packaging==25.0
pandas==2.3.3
pephubclient==0.4.5
peppy==0.40.8
phonemizer==3.3.0
pillow==12.0.0
piper-tts==1.3.0
pipestat==0.12.2
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
PyAudio==0.2.14
pycaw==20251023
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.2.0
rpds-py==0.30.0
segments==2.3.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
SQLAlchemy==2.0.45
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubiquerg==0.8.2
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
yacman==0.9.4
yarl==1.22.0
zstandard==0.25.0